Move resign endpoint to assigment controller (#2106)
