+++ packages/apps/fortune/exchange-oracle/server/src/modules/assignment/assignment.controller.spec.ts
@@ -3,7 +3,7 @@ import { Test } from '@nestjs/testing';
 import { AssignmentController } from './assignment.controller';
 import { AssignmentService } from './assignment.service';
 import { RequestWithUser } from '../../common/types/jwt';
-import { GetAssignmentsDto, CreateAssignmentDto } from './assignment.dto';
+import { GetAssignmentsDto, CreateAssignmentDto, ResignDto } from './assignment.dto';
 import { AssignmentStatus, JobType } from '../../common/enums/job';
 import { MOCK_EXCHANGE_ORACLE } from '../../../test/constants';
 
@@ -84,6 +84,27 @@ describe('assignmentController', () => {
       expect(assignmentService.createAssignment).toHaveBeenCalledWith(body, {
         address: userAddress,
       });
+    });
+  });
+
+  describe('resignJob', () => {
+    it('should call jobService.resignJob', async () => {
+      const assignmentId = 123;
+      const resignJobDto: ResignDto = {
+        assignmentId,
+      };
+
+      jest.spyOn(assignmentService, 'resign').mockResolvedValue();
+
+      await assignmentController.resignJob(
+        { user: { address: userAddress } } as RequestWithUser,
+        resignJobDto,
+      );
+
+      expect(assignmentService.resign).toHaveBeenCalledWith(
+        resignJobDto.assignmentId,
+        userAddress,
+      );
     });
   });
 });
