+++ packages/apps/fortune/exchange-oracle/server/src/modules/assignment/assignment.controller.ts
@@ -20,6 +20,7 @@ import {
   AssignmentDto,
   CreateAssignmentDto,
   GetAssignmentsDto,
+  ResignDto,
 } from './assignment.dto';
 import { RequestWithUser } from '../../common/types/jwt';
 import { PageDto } from '../../common/pagination/pagination.dto';
@@ -94,5 +95,34 @@ export class AssignmentController {
       req.user.reputationNetwork,
       serverUrl,
     );
+  }
+
+  @ApiOperation({
+    summary: 'Resign Job',
+    description: 'Endpoint to resign from a job.',
+  })
+  @ApiBearerAuth()
+  @ApiBody({
+    description: 'Details required to resign from the job.',
+    type: ResignDto,
+  })
+  @ApiResponse({
+    status: 200,
+    description: 'Job resigned successfully.',
+  })
+  @ApiResponse({
+    status: 400,
+    description: 'Bad Request. Invalid input parameters.',
+  })
+  @ApiResponse({
+    status: 401,
+    description: 'Unauthorized. Missing or invalid credentials.',
+  })
+  @Post('resign')
+  resignJob(
+    @Request() req: RequestWithUser,
+    @Body() body: ResignDto,
+  ): Promise<void> {
+    return this.assignmentService.resign(body.assignmentId, req.user.address);
   }
 }
