+++ packages/apps/fortune/exchange-oracle/server/src/modules/assignment/assignment.dto.ts
@@ -116,3 +116,9 @@ export class AssignmentDto {
     this.expiresAt = expiresAt;
   }
 }
+
+export class ResignDto {
+  @ApiProperty({ name: 'assignment_id' })
+  @IsNumber()
+  public assignmentId: number;
+}
