+++ packages/apps/fortune/exchange-oracle/server/src/modules/job/job.dto.ts
@@ -36,12 +36,6 @@ export class SolveJobDto {
   public solution: string;
 }
 
-export class ResignJobDto {
-  @ApiProperty({ name: 'assignment_id' })
-  @IsNumber()
-  public assignmentId: number;
-}
-
 export class GetJobsDto extends PageOptionsDto {
   @ApiPropertyOptional({
     name: 'sort_field',
